feat(sentry): Refactored and added slices in Sentry store
