+++ .changeset/famous-dryers-sing.md
@@ -1,0 +1,5 @@
+---
+'@spotlightjs/overlay': minor
+---
+
+Sentry Store Slices
