+++ packages/overlay/src/integrations/sentry/components/insights/Profiles.tsx
@@ -4,8 +4,8 @@ import { ReactComponent as SortDown } from '~/assets/sortDown.svg';
 import classNames from '~/lib/classNames';
 import Table from '~/ui/Table';
 import { AGGREGATE_CALL_PROFILES_SORT_KEYS, AGGREGATE_PROFILES_HEADERS } from '../../constants';
-import useSentryStore from '../../data/sentryStore';
 import useSort from '../../hooks/useSort';
+import useSentryStore from '../../store';
 import type { AggregateCallData } from '../../types';
 import { getFormattedDuration, getSpanDurationClassName } from '../../utils/duration';
 import { TimeBar } from '../shared/TimeBar';
