+++ packages/overlay/src/integrations/sentry/components/insights/envelopes/EnvelopeList.tsx
@@ -4,9 +4,9 @@ import CardList from '~/components/CardList';
 import TimeSince from '~/components/TimeSince';
 import classNames from '~/lib/classNames';
 import Badge from '~/ui/Badge';
-import useSentryStore from '../../../data/sentryStore';
 import { useSentryEnvelopes } from '../../../data/useSentryEnvelopes';
 import { useSentryHelpers } from '../../../data/useSentryHelpers';
+import useSentryStore from '../../../store';
 import { sdkToPlatform } from '../../../utils/sdkToPlatform';
 import { truncateId } from '../../../utils/text';
 import PlatformIcon from '../../shared/PlatformIcon';
