+++ packages/overlay/src/integrations/sentry/data/useSentryHelpers.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
+import useSentryStore from '../store';
 import { SentryEventsContext } from './sentryEventsContext';
-import useSentryStore from './sentryStore';
 
 export const useSentryHelpers = () => {
   useContext(SentryEventsContext);
