+++ packages/overlay/src/integrations/sentry/data/useSentrySdks.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
+import useSentryStore from '../store';
 import { SentryEventsContext } from './sentryEventsContext';
-import useSentryStore from './sentryStore';
 
 export const useSentrySdks = () => {
   useContext(SentryEventsContext);
